Allows 'two-wheel-drive' rovers (i.e. two motorized wheels, two non-motorized landing gear) to be automated.
Also no longer depends on ModuleWheelDamage to say if a wheel is touching the ground or not.

Also allows rovers with broken wheels to be automated, *if* there are enough non-broken wheels still touching the ground (i.e. 4, with at least 2 motorized).  This can be changed by moving the isGrounded check after the damage check.